refactor: extract recording controls from App
